backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
django-cors-headers>=4.3
psycopg2-binary>=2.9
channels>=4.0
channels-redis>=4.1
daphne>=4.0
djangorestframework-simplejwt>=5.3
Pillow>=10.0

# Voice Control & Keyword Spotting
tensorflow>=2.15
tensorflow-io>=0.36
numpy>=1.24
scipy>=1.11
